chore: fix clippy warnings

+++ crates/kernel/src/utils/bits.rs
@@ -1,5 +1,5 @@
 pub fn get_bit(value: u64, bit: usize) -> u64 {
-    let mask = (1 as u64) << bit;
+    let mask = 1_u64 << bit;
     let masked = value & mask;
     (masked > 0) as u64
 }
@@ -7,7 +7,7 @@ pub fn get_bit(value: u64, bit: usize) -> u64 {
 pub fn get_bits(value: u64, low_bit: usize, high_bit: usize) -> u64 {
     let mut ret: u64 = 0;
     for bit in low_bit..high_bit {
-        ret |= (get_bit(value, bit) as u64) << (bit - low_bit);
+        ret |= get_bit(value, bit) << (bit - low_bit);
     }
 
     ret
